refactor: move signup factories to a folder
