+++ src/main/factories/signup/SignUpControllerFactory.ts
@@ -1,10 +1,10 @@
-import SignUpController from '../../presentation/controllers/signup/SignUpController'
-import DbAddAccount from '../../data/usecases/add-account/DbAddAccount'
-import BcryptAdapter from '../../infra/criptography/BcryptAdapter'
-import AccountMongoRepository from '../../infra/db/mongodb/account-repository/AccountMongoRepository'
-import LogControllerDecorator from '../decorators/LogControllerDecorator'
-import Controller from '../../presentation/protocols/Controller'
-import LogMongoRepository from '../../infra/db/mongodb/log-repository/LogMongoRepository'
+import SignUpController from '../../../presentation/controllers/signup/SignUpController'
+import DbAddAccount from '../../../data/usecases/add-account/DbAddAccount'
+import BcryptAdapter from '../../../infra/criptography/BcryptAdapter'
+import AccountMongoRepository from '../../../infra/db/mongodb/account-repository/AccountMongoRepository'
+import LogControllerDecorator from '../../decorators/LogControllerDecorator'
+import Controller from '../../../presentation/protocols/Controller'
+import LogMongoRepository from '../../../infra/db/mongodb/log-repository/LogMongoRepository'
 import SignUpValidationCompositeFactory from './SignUpValidationCompositeFactory'
 
 const SignUpControllerFactory = {
