+++ src/main/factories/signup/SignUpValidationCompositeFactory.spec.ts
@@ -1,11 +1,11 @@
-import ValidationComposite from '../../presentation/helpers/validators/ValidationComposite'
+import ValidationComposite from '../../../presentation/helpers/validators/ValidationComposite'
 import SignUpValidationCompositeFactory from './SignUpValidationCompositeFactory'
-import RequiredFieldValidation from '../../presentation/helpers/validators/RequiredFieldValidation'
-import Validation from '../../presentation/helpers/validators/Validation'
-import EmailValidation from '../../presentation/helpers/validators/EmailValidation'
-import EmailValidator from '../../presentation/protocols/EmailValidator'
+import RequiredFieldValidation from '../../../presentation/helpers/validators/RequiredFieldValidation'
+import Validation from '../../../presentation/helpers/validators/Validation'
+import EmailValidation from '../../../presentation/helpers/validators/EmailValidation'
+import EmailValidator from '../../../presentation/protocols/EmailValidator'
 
-jest.mock('../../presentation/helpers/validators/ValidationComposite')
+jest.mock('../../../presentation/helpers/validators/ValidationComposite')
 
 class EmailValidatorStub implements EmailValidator {
   isValid (_: string): boolean {
