+++ src/main/factories/signup/SignUpValidationCompositeFactory.ts
@@ -1,8 +1,8 @@
-import Validation from '../../presentation/helpers/validators/Validation'
-import ValidationComposite from '../../presentation/helpers/validators/ValidationComposite'
-import RequiredFieldValidation from '../../presentation/helpers/validators/RequiredFieldValidation'
-import EmailValidation from '../../presentation/helpers/validators/EmailValidation'
-import EmailValidatorAdapter from '../../utils/EmailValidatorAdapter'
+import Validation from '../../../presentation/helpers/validators/Validation'
+import ValidationComposite from '../../../presentation/helpers/validators/ValidationComposite'
+import RequiredFieldValidation from '../../../presentation/helpers/validators/RequiredFieldValidation'
+import EmailValidation from '../../../presentation/helpers/validators/EmailValidation'
+import EmailValidatorAdapter from '../../../utils/EmailValidatorAdapter'
 
 const SignUpValidationCompositeFactory = {
   makeSignUpValidationComposite (): Validation {
